feat: use native parseArgs in the optimize script

+++ app/src/scripts/optimize/optimize.ts
@@ -1,4 +1,4 @@
-import { join } from 'node:path'
+import { resolve } from 'node:path'
 import dotenv from 'dotenv'
 
 import CloudinaryImage from '@/lib/image.js'
@@ -6,28 +6,53 @@ import { handleLogError } from '@/utils/helpers.js'
 
 dotenv.config()
 
-const main = async () => {
-  try {
-    const dir = join(process.cwd(), 'assets')
-    const file = join(dir, 'mecha_poster_02.png')
+const log = (str: string) => {
+  console.log(`[OPTIMIZE] ${str}`)
+}
 
+export const optimize = async (args: Record<string, string | boolean>) => {
+  const {
+    file = '',
+    assetfolder = 'image-optimizer',
+    outputFolder = '',
+    tags = '',
+    width = 800,
+    upload = false,
+    deleteAfter = false,
+  } = args
+
+  const filePath = String(file)
+  let outDir = null
+
+  if (outputFolder) {
+    outDir = resolve(String(outputFolder))
+  }
+
+  log(`Input file:\n${filePath}`)
+  log(`Output folder: ${outDir}`)
+  log(`Cloudinary asset folder: ${assetfolder}`)
+  log(`Tags: ${tags}`)
+  log(`Image width: ${width}`)
+  log(` -upload? ${upload}`)
+  log(` -delete cloud file? ${deleteAfter}\n`)
+
+  try {
     const image = new CloudinaryImage({
-      localFile: file,
-      cloudinaryAssetFolder: 'image-optimizer',
+      localFile: filePath,
+      cloudinaryAssetFolder: String(assetfolder),
+      ...(outDir && { localDestination: outDir }),
     })
 
-    await image.upload('gundam,seed,mobile suit')
-    await image.optimize()
-    await image.delete()
+    if (upload) {
+      await image.upload(String(tags))
+    }
+
+    await image.optimize(Number(width))
+
+    if (deleteAfter) {
+      await image.delete()
+    }
   } catch (err) {
     handleLogError(err, 'ERROR')
   }
-}
-
-if (process.env.IS_DOCKER === 'true') {
-  setTimeout(() => {
-    main()
-  }, 5000)
-} else {
-  main()
 }
